requirements.txt
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autocommand==2.2.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cheroot==9.0.0
CherryPy==18.8.0
colorama==0.4.6
confusable-homoglyphs==3.2.0
cryptography==38.0.3
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.4
django-3-jet==1.0.8
django-celery-beat==2.4.0
django-ckeditor==6.5.1
django-constance==2.9.1
django-filter==22.1
django-hashedfilenamestorage==2.4.1
django-js-asset==2.0.0
django-picklefield==3.1
django-ratelimit==3.0.1
django-redis==5.2.0
django-timezone-field==5.0
django-widget-tweaks==1.4.12
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.0.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.2
httpx==0.23.1
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
inflect==6.0.2
iniconfig==1.1.1
ipaddress==1.0.23
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.2.0
jaraco.functools==3.5.2
jaraco.text==3.11.0
jsonpointer==2.3
loguru==0.6.0
more-itertools==9.0.0
oauthlib==3.2.2
packaging==21.3
passlib==1.7.4
phonenumbers==8.13.0
Pillow==9.3.0
pluggy==1.0.0
portend==3.1.0
priority==2.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
pyQiwiP2P==2.0.6
pytest==7.2.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
response-report==1.1.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.22.0
tempora==5.1.0
toml==0.10.2
tomli==2.0.1
toposort==1.7
typing==3.7.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
validators==0.20.0
win32-setctime==1.1.0
wsproto==1.2.0
YooMoney==0.1.0
zc.lockfile==2.0
